+++ python/ambers/__init__.py
@@ -5,11 +5,15 @@ from ambers._containers import SavFile
 from ambers._io import read_sav, read_sav_meta, scan_sav, write_sav
 from ambers._missing import apply_missing
 from ambers._transforms import apply_labels
+from ambers._validate import ValidationIssue, ValidationReport, validate
 
 __all__ = [
     "SavFile",
     "apply_labels",
     "apply_missing",
+    "validate",
+    "ValidationReport",
+    "ValidationIssue",
     "read_sav",
     "read_sav_meta",
     "scan_sav",